create new post-item comp
