+++ src/components/post-listing/post-listing.component.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import { Link, useStaticQuery, graphql } from 'gatsby';
-import CardActionArea from '@material-ui/core/CardActionArea';
-import CardContent from '@material-ui/core/CardContent';
-import Img from 'gatsby-image';
-import { StyledContainer, StyledCard } from './post-listing.style';
-import placeHolder from '../../img/placeholder.png';
+import { StyledContainer } from './post-listing.style';
+import PostItem from '../post-item/post-item.component';
 
 const PostListing = () => {
   const data = useStaticQuery(graphql`
@@ -43,7 +40,7 @@ const PostListing = () => {
   const getPostList = () => {
     const postEdges = data.allMarkdownRemark.edges;
     const postList = [];
-    postEdges.forEach(postEdge => {
+    postEdges.forEach((postEdge) => {
       postList.push({
         path: postEdge.node.fields.slug,
         tags: postEdge.node.frontmatter.tags,
@@ -52,7 +49,7 @@ const PostListing = () => {
         date: postEdge.node.fields.date,
         excerpt: postEdge.node.excerpt,
         timeToRead: postEdge.node.timeToRead,
-        id: postEdge.node.id
+        id: postEdge.node.id,
       });
     });
     return postList;
@@ -62,7 +59,7 @@ const PostListing = () => {
 
   return (
     <StyledContainer>
-      {postList.map(post => {
+      {postList.map((post) => {
         let cover;
         if (post.cover) {
           cover = post.cover.childImageSharp.fluid;
@@ -70,22 +67,7 @@ const PostListing = () => {
 
         return (
           <Link key={post.id} className='styledLink' to={post.path}>
-            <StyledCard>
-              <CardActionArea className='cardActionArea'>
-                {cover ? (
-                  <Img fluid={cover} />
-                ) : (
-                  <div
-                    className='cardActionArea__placeholder'
-                    style={{ backgroundImage: `url(${placeHolder})` }}
-                  />
-                )}
-                <CardContent className='cardContent'>
-                  <h2>{post.title}</h2>
-                  <p>{post.excerpt}</p>
-                </CardContent>
-              </CardActionArea>
-            </StyledCard>
+            <PostItem cover={cover} title={post.title} excerpt={post.excerpt} />
           </Link>
         );
       })}
